Cleanup of 2015 Day 19
So I ended up using "lark" to parse the grammar in part two. This is
sort of cheating, but I'm not really that interested in implementing the
LALR or CYK algorithm myself at this point. It ended up being quite
slow to implement, so I think going for something simpler during the
contest.

My input grammar could be parsed with LALR, but some of the examples
couldn't. For good measure, I added a fallback to the earley parser if
the LALR parser doesn't work.

I've learned two things through the last few exercises though. I've got
to be more comfertable using the regex module in python. For many of the
challenges it will be the fastest way.

The second is that just winging it will work in most cases