2015 Day 20
This one was kind of meh to solve, but I think it will end up being
interesting to see what a proper solution here looks like.

I started trying part one with just doing mod for all numbers below the
current one. This is too slow for the high number, so I knew I needed to
go to some sort of sieve. The Sieve of Eratosthenes is one of the few
sieves that I remember, so I wanted to go with that one.

For this to work, I need to establish an upper limit. I started thinking
about least common denominators and so forth, but in the end, it ended
up begin rather easy to get a worst case. Eventually there will be an
elf with LIMIT/10 as their number. The first house they visits, will get
at least LIMIT number of points.

The whole 10 times their number thing is a smokescreen, you can do the
entire task without thinking about that. The answer is going to be a sum
of numbers, which are all multiplied by 10, so we can factor the
multiplication out till the end. We can then just divide the LIMIT by 10
and use that throughout instead.

Ended up using "tqdm" to check if this was fast enough. It seemed it
would take to long in the beginning, but as the step size become bigger,
it speeds up considerably.

The answer was right on the first try!

Happy with my conversion on the second part, I just changed some numbers
and added a limit to the steps. This actually brought the iteration time
even further!

Lessons for next time I think is just to start doing a "half-smart"
solution right away.

Part one: 00:18:44
Part two: 00:01:55