Cleanup of 2015 Day 20
This ended up being very neat once I cleaned it up. It's still slow in
CPython, but with pypy it's near instantaneous! I'll look at the answers
in the thread after this.

One neat trick I learned in this challenge is that you can limit an
iterator by using "zip" and "range". "zip" will only iterate as long as
there is data from all inputs, so it terminates on whatever is shorter.
This is a really neat way to put limits on iterators!