2015 Day 21
This was a fun one, though a bit easy!

The first part I'm not sure what I spent the most time on, I think it
must have been the parsing of the weapons and so on. I headed down the
right path here though, with just writing the nested loops directly
instead of trying to be clever. I also did a couple of other things that
I'm very pleased with:

I didn't bother reading in the boss stats from the input, rather I
hardcoded it into my code. This saved unneeded complexity and can be done
later.

I inspected the input and saw that it was kind and so the parsing became
quite easy. I wasn't sure if `str.split` collapsed sequential
white space, but it did, so it worked out great.

I read the text carefully and so spotted the fact that you always
inflict one point of damage.

I wrote the loops directly instead of trying to be cute. This is
especially true of the rings, where I saw that none of them shared the
same stats, so I didn't have to fiddle with indexes, I could just do a
direct comparison.

My conversion to part two I'm also very pleased with, it was essentially
just changing out two conditionals.

I'll clean this up a bit and as always I'm looking forward to seeing if
there are any clever ways to solve this on the forums.

Part one: 00:17:10
Part two: 00:01:28