Cleanup of 2015 Day 21
This reduced a bit, but overall very much the same. I looked on the
forums and really the only difference between what I did and what was
there is that some peopled added a zero armour and zero rings, to avoid
having to test if rings was there or not. This was clever, but it
doesn't really add much over my method.

This would have been difficult if there was many more items, then it
would have been sort of a knapsack problem I think.