2015 Day 22
This was an interesting one! It also took me a long time!

There were many details here, so just getting the logic right was a
major hurdle. I started out trying to make it generic, but abandoned it
in favor of doing it the fast way. I ended up hardcoding the spells and
the effects.

When I finally got it going, it was way too slow. This is essentially a
graph search and I have some experience with optimizing those. My first
idea was to start up and end the program a couple of times to see if
there was a common place where we were using time. This turned out to be
a deep copy operation I was doing on the effect dict.

I rewrote it to a tuple and that made it a bit faster, but not much. My
next idea then was to introduce a global minimum. Whenever we actually
win we can check what the current cost down here is and then set this as
a new golden standard. Any branch of the tree which goes above this can
be pruned.

Writing this took a little bit, but it made the answer instantaneous.
Adapting the code to part two was trivial. This one is going to take a
lot of cleaning up, large parts of my code is not needed any more, as
the single global minimal variable is all that is needed, my functions
don't actually need to return anything. This makes the code a lot
simpler.

Next time when doing DFS I should remember this pruning technique and
that it makes the result code a lot simpler.

Part one: 00:49:39
Part two: 00:01:25