2015 Day 4
I'm redoing the ones I've already done, just to have them all in python.
This one was not really that fun to program, it's just a bruteforce. It
ends up being quite slow in python, but fast enough to be doable.

Learned about the hashlib library today. The timings aren't really all
that important here, as I've done the tasks before, but if I can pick up
some skills then it is useful. It would also be nice to have done all
the tasks in python, to have a reference implementation to compare other
languages that I might end up using.

Part one: 00:04:40
Part two: 00:00:32