2015 Day 24
A balancing task!

I think I did passable on the first part. I reached for "more_itertools"
immediately and found "set_partitions" and "powerset" right away, but
just using "set_partitions" was too slow.

Then I realized that I could prune a lot of the possibilities since the
first group is always smallest. From there it was just a question of
using "combinations" with stepped execution to filter out the first
level. I could then use partitions for the rest of the problem.

This was fast to output one answer, which I went for, and which is the
correct answer.

In part two I tried to just add one more to the "set_partitions"
strategy of part one, but this didn't work. I ended up repeating the
pattern from part one with the "combinations" function, but I can set
the length to the same as "a", as I know it cannot be shorter, or it
would be "a".

This also found a solution quickly, but didn't terminate quickly. After
having submitted I added a rule which prunes out alternatives that have
a higher QE score than what we have found. This finishes quite quickly,
but the first solution we find is the correct one each time.

I realized the reason for this is that since we are iterating over the
possible "a" sets in lexicographical order, the first solution found
will always have the smallest product. We can quite once we find the
first solution. I'll implement this in the cleanup.

Overall happy with my performance, never really did nothing, I was
always coding.

Part one: 00:17:32
Part two: 00:04:12