2015 Day 25
Kinda anticlimactic :/

This one was all about understanding the problem and coming up with a
way to iterate all the way to the answer. I quickly came up with a
solution there, and spent some time debugging it.

I'm sure there is some cool math way to calculate exactly which number
in the series a given row and column will be. If you can calculate that,
we don't need to do the position calculations, we can simply do the key
algorithm until we reach that number.

Part one: 00:09:12
Part two: 00:00:09