Cleanup of 2015 Day 25
Realized two things from reading the reddit thread! The first is that
there is in fact a way to compute the index directly, but I haven't
worked that out yet.

The second part is that this is just modular exponentiation and I can
simply wait with doing any complicated calculations til the end. This
should make this a O(1) once I figure out the index calculation!

I think this task is greater than it first appears ;)