2016 Day 04
Fun task!

Wrote the regex quite quickly, but should have been faster. Used count
and so on pretty effectivly in part one, had a bit of trouble because I
didn't remove the dashes, but when I did it worked.

Part two was the first task where there has been real ambigouity I feel
like. I was slow to implement the shift, and I implement it in a stupid
way, but which was easy to code. When I'm done with this, I'm going to
write it in a proper way, which calculates the value directly.

To get the answer I grepped for pole and found what I have now. Don't
know if this is always true, but it probably is.

Part one: 00:10:32
Part two: 00:06:15