2016 Day 5
Another MD5 challenge, but today I messed up reading the instructions
properly and so it took some time to get started. This ones also takes a
lot longer to compute than previous years. I might buckle down and make
this one use parallel programming to speed it up, we will see.

Part two was also not so good, I failed to read all the instructions
carefully again. Not pleased with the times overall, but I've had a real
shitty evening and so this is a fitting end.

Part one: 00:08:17
Part two: 00:06:04