Cleanup of 2016 Day 7
This cleaned up very nicely! It runs much faster and it's quite easy to
understand. Things I learned about this task is that whenever the
patterns aren't all that regular, it's worth doing it by hand quickly.

I also learned that python will create a new new group before the first
identifier in a "str.split()" scenario, so I could use it here to know
in and out groups.

Overall, very pleased with the results, learned something here :)