2016 Day 9
Quite quick adaption of the original I would say.

This was a really nice task I think. The clue quickly is to realize that
you don't need to build the string, but rather just the length. This
took me a little bit, but my solution was quickly adoptable to just
couting the length.

I did the right thing and just did it stupid simple, iterating through
the string. In hindsight I should have just kept shortening the string,
to make the indexing a bit easier, because I spent quite some time on
that. But once I got it working it was fast. In the cleanup, I'll
explore this alternative.

The conversion for part two was quite quick, I just needed to pull my
solving step into it's own function. This might be a good habbit to get
into in general?

I wonder what the complexity for this task is? It seems like it could be
O(n), given we only progress over the string ones, but for part two, we
have to parse each substring.

Part one: 00:11:10
Part two: 00:03:27