2016 Day 11
Jeeesussss! First really tough problem in 2016, I had to pull out all
the stops to get this one.

I spent quite some time understanding the rule set, but it pales in
comparison to the time I spent trying to write a brute force solution
for this.

A lot of time was spent making the classes and getting that aspect to
work. Then I had to spend a lot of time on debugging some issues where
invalid states would be created, because the "Generator" and "Chips"
classes where too alike I think? Adding a second parameter, and making
those values unique fixed it, but Jesus that was some bug to try to
track down.

For part one I implemented this as a BFS and then as a DFS. The DFS
worked. I didn't wait for it to try all the possibilities, when it got
to 33, I simply entered the answer.

For part two, I tried the same tactic, but it didn't work. What I ended
up doing was introducing the concept of a "score" heuristic, which is
a rough approximation about how close we are to the solution. With this
metric introduced I was able to find a solution with 81 moves. This
worked it's way downward, but not really fast enough.

While this was running I would try to tweak the score heuristic and see
if that made a difference. Sometimes it would give a big leap in the
right direction, finding a new minimum that was smaller than the current
running process. Once that happened, I would kill the other process and
try to tweak the score again.

Finally it found a minimum at 65, which went down to 61 and then 59. I
tried each of these answers, but they where all wrong. But by now, I've
discovered that all answers are odd numbers. So I tried 57 later and it
turned out to be correct.

The second program is still running. I'm going to leave it running, to
see if it eventually finds the correct solution.

I'm so stocked to try to puzzle out the real solution here. My guess is
that there is a rather simple tactic that will always be optimal. I
don't think a general graph search is the intended solution for this
problem. If it is, then I'll be very interested in seeing how it looks!

Rather pleased that I managed to solve this one, even if it took 4
hours. This is the sort of tasks that really lead to learning, so I'm
going to study the solution here!

Part one: 02:59:00
Part two: 01:05:18