2015 Day 6
This was a rather easy task, it resembles a lot of other tasks in AoC. I
used numpy to solve this one and it turned out to be quite a fast
solution to the problem.

I spent a little bit looking for the "clip" function for negative
operations, but other than that it went swimmingly.

Part one: 00:06:10
Part two: 00:03:56