Cleanup of 2016 Day 11
Well, this is the strangest thing ever. I was playing around with this,
trying to make my current solution faster. I realized that I could cut a
lot of the states by just looking at the relative position of the pairs
of generator and microchips and then sorting that list. That way we are
not doing all branches to the goal, if there are two identical paths to
the goal.

This didn't speed anything up noticabily however, when I changed the
"__eq__" function, but the second I changed the "__hash__" function,
things just went freakisly fast. I'm not sure why this is, I'm guessing
it now does a lot less comparisons? Either way it brough both solutions
down to under 10 seconds. This is remarkable and something I'm going to
investigate further.

I will also be cleaning this task up and writing a C++ implementation
also. I want to attempt to encode the state as a single 64bit integer.