2015 Day 7
This is quite a fun day, made a lot easier by numpy having the uint16
type.

I'm ok with my time. I assume that there is a lot of room for
optimisation here, but it's ok fast as it is now. I'll probably look
into removing numpy from this, to make it into pure python.

Part one: 00:11:08
Part two: 00:01:22