2016 Day 12
More ISA tasks!

This one was quite simple, but I struggled with a bug where I didn't
increment the jump counter if the "jnz" instruction didn't perform the
jump. Should have caught this earlier, but such is the way.

Part two takes quite some time, so I'm thinking of ways to optimize this
away. My current idea is to try to convert the instruction set into a
program and then just eval that. Another thing I can try is to parse the
program into instructions once and then just execute them. The last is
to try to do loop detection of some sort and try to optimize this away.

Another thing to do would be symbolic execution of some sort. I'd be
interested in seeing what the various inputs looks like.

Having quite a bad day, so this was a nice little treat.

Part one: 00:10:29
Part two: 00:00:42