Optimize 2016 Day 12
So I did create the transpilation like I discussed in the previous
commit message. Modern compilers really are amazing, clang is able to
deduct the answer during the compiler phase, it doesn't even have to run
a single loop.

I wish I was able to code the loop detection in python, to avoid the
dependency on "clang" or "gcc", but it would be more work than it's
worth.