2016 Day 13
This one is a pure graph problem. I started doing something akin to A*,
but realized that it was not optimal. Switched it out with a standard
BFS and that worked just fine. I was surprised that this worked as well,
I was expecting more resistance.

My conversion to part 2 was quick enough, but I managed to set the
boundary wrong, 49 instead of 50. This was due to a bit of mental slip,
but once it was corrected it worked straight away.

Part one: 00:15:38
Part two: 00:03:09