2016 Day 14
This was a clusterfuck from my side, but I did do some small things
right. To begin with I saw that it was the annual MD5 challenge and so I
quickly turned to "hashlib". One thing I did do right, which was quite
nice is that I made a simple program to generate a list of hashes for
the given input and then used that in the main program. This allowed me
to experiment pretty quickly, without having to do the md5 stuff again.

This probably didn't speed it up all that much, but it was a nice little
exercise. Figuring out the logic which I should do this by was not as
nice. I ended up doing a lot of hacky things and this will be really
important to clean up. I think the right thing to do here would have
been to generate 1000 extra hashes for each potential key and just check
it right away, instead of trying to keep track of the potential keys as
I did.

Once that was cracked though, I ended up getting part two quite quickly,
but not as quickly as I would have liked. It took me some time to
realize if they where doing 2016 rounds of hashing then outputting one
or outputting 2017 hashes per round.

I rewrote the logic here to use the hashlib directly, and this is what
took the 6 minutes mainly.

Overall not very pleased with this, but I learned some things. I think I
will learn a lot more from the cleanup.

Part one: 00:33:34
Part two: 00:06:35