2016 Day 15
This was a rather good one, because it reminded me of the stuff I did
for 2020 Day 13. It's the Chinese remainder theorem. Got it out pretty
quickly, but it took some time to grok the task, I found the description
a bit confusing.

Part two was an easy conversion.

Part one: 00:14:22
Part two: 00:01:47