2016 Day 16
Ok, so this one was ok, I spent a bit longer than I would have liked
parsing the instructions. The implementation itself was easy in python,
I just went for the simplest solution, which was to use strings and to
compute the whole thing. I think this can be done a whole lot faster,
but for now that was ok.

Something which was not so ok, was how big I fucked this up on the
second part, due to clipboard messing around. I would copy various
things and not think about it, which is why my first answer was wrong.
Then I would get the correct answer, and copy that, and try to resubmit,
but that would again not work because I was too quick.

Then I would run the program again, because I'm stupid like that, but
now my original input is not in my clipboard so I get the wrong answer.
Then I try to submit that and get the wrong answer and realize the
mistake in my ways. Had to sit for a minute waiting to submit.

Lessons from this one: Make sure to have the puzzle input copied into
the clipboard and be patient when trying to resubmit.

Part one: 00:08:48
Part two: 00:02:15