2016 Day 18
This was a cool day and I am very pleased with my time here! Used a
trick to avoid tricky edge calculations, and just padded the input
array. Code is slow for part two, but it got the job done. I am sure
there is some faster way to do this, but for now this is fast enough.

Parsed the input quite quickly. Again very pleased.

Part one: 00:07:02
Part two: 00:00:45