Cleanup of 2016 Day 19
So the Reddit thread was very useful here. Turns out that the first task
is a known problem called the "Josephus problem". There is a simple
solution to it. The second problem there is no solution for formally,
but there are ways to speed up what I did.

The first, which is the one I might have had a change to do is to
maintain two queues of the two sides, and juggle people between these.
This is quite fast and works great.

The second is to observe the output for a pattern and then apply that
pattern. I've included both solutions in my files, as I feel like they
are not totally of my making.