2016 Day 20
This one was quite fun, though I am displeased with my time on part 2.
Most of my time on part two was spent figuring out how to use the
intervaltree package from pip.

I am glad I reached for a library for this and that I recognized it.
I'll add the intervaltree to my template and use it in further
challenges.

Part one: 00:09:53
Part two: 00:21:19