2016 Day 21
This was a rather tedious task for part 1. It was mostly just parsing
the rules. For part two I ended up doing something a bit stupid. I spent
quite some time thinking about if it was possible to just view the
hashing processes as a static process and just work backwards. I gave up
on this after little while and decided to just bruteforce it. This works
quick enough.

It will be interesting to study this one some more. I have a nice
feeling that z3 should be able to be used here here and give the answer
that we want. It's been long since I've used z3, so it will be a good
exercise.

Part one: 00:18:32
Part two: 00:06:48