2016 Day 22 Part 1
This task is a hard hitter, I really don't know how to do part two! I've
managed to solve the test input, by doing a semi A* search, but it's not
going to cut it for part two.

I've cheated a bit and I know that the answer is above a 100 and so I've
given up on the naive approach. My only clue now is that it might be
some sort of simplification that can be made about the grid. Maybe this
is really like one of those puzzles where you have to slide things
around to make the picture appear.

Part one: 00:11:00
Part two: 01:38:33 (so far)