2016 Day 23
Another ISA task! Spent a bit of time on part one getting my old code
which was more condusive to doing this sort of task than the polished
one I had made. Then I spent the next 20 minutes not realizing that we
needed to initalize a with 7. I really have to read the instructions
more.

Part two seems to imply that you cannot just run the same interpreter.
I beg to differ! It didn't take all that long and it's clear it's doing
some sort of multiplication. I'm not going to look more into this right
now, but it needs to be investigated at a later date.

Part one: 00:32:50
Part two: 00:03:33