2016 Day 24
Pleased with this one! I wrote the basic A* search quite quickly and
then created a function to find the optimal combinations. I'm sure this
wouldn't have worked for a larger grid, but for now, just comparing all
alternatives was good enough. I could have pruned it more I guess, but
there was little point since it finished quickly.

The conversion in part was quite quick, quite proud of that!

Part one: 00:23:04
Part two: 00:01:09