2016 Day 25
Well, it was sort of an easy end to the whole thing, but it was nice
that it reused the assembunny ISA. I did a dirty hack to get it sort of
right and the answer popped right out. I haven't solved day 22 yet, but
other than that I'm finished with 2016.

Part one: 00:07:30