2017 Day 7
This was a good day! A novel task for part two. I thought it was going
to be quite hard, but once I realized that there would always be just
one branch to go down, it was quite easy. The code ended up being quite
messy, but I was going for speed.

Part one: 00:04:09
Part two: 00:20:45