2017 Day 9
I always like parsers, so this was an ok day. I spent some time figuring
out how I was going to count the groups, but I figured out quite a
simple way to do it.

The second part was easy to solve with my solution, so that was nice.
Not the most impressive day by me, but definitively an ok time.

Part one: 00:13:10
Part two: 00:01:24