2017 Day 11
This is familiar to a previous problem, where there was also a hex grid.
I remembered from that you can express it as a complex number. I went to
my main resource for this[1] and quickly got the formula. I did however,
fuck it up with the distance calculation. I swapped an imaginary for a
real part in one of the operations and that was it.

Once that was sorted it went really quick. Disappointed with my time in
this one, but a good reminder that the guy likes his hex grids.

[1]: https://www.redblobgames.com/grids/hexagons/

Part one: 00:14:35
Part two: 00:01:32