2017 Day 13
So this one was quite cool. It smelled of the Chinese remainder theorem
when I saw it, but the up and down motion was throwing me off, so I went
with a naive implementation instead in the beginning. This got me
through part one, but part two wanted to know how to get through.

I knew straight away that the naive solution wouldn't work, but I did a
simple solution and just set that to work in the background. Meanwhile I
was trying to figure out how to convert it to a simple modulus problem.
I actually did figure that out, but I was doing the testing in a silly
way.

Once I figured that out, I started a background thread with the faster
implementation and was about to go back and figure out how to do this
smartly, when I saw that it had finished. I'm sure there is a proper way
to do this, but right now, this will hold :)

Part one: 00:10:45
Part two: 00:12:53