2017 Day 14
Combo day! This was a combination of day 12 and day 10! I saw this quite
quickly and was able to get it setup for part two, but again knot hash
fucked me.

I put a zero padding of 4 on the output of the string instead of a
padding of 8, which is what you need for a two digit hexadecimal string.
This worked for part one, since we where only counting the zeroes, but
for part two, the alignment became off. I should have thought of this
earlier, but spent a lot of time debugging this.

Once that was fixed, it worked straight away.

Part one: 00:05:02
Part two: 00:14:06