2017 Day 16
This was a fun day! I got part one quite quickly, but quickly found that
part two was going to require some different sort of tricks. I added a
cache and recorded the number of cache hits that I got. I saw that after
a while there was no more cache misses, which led me to believe that we
had hit a loop.

This was indeed the case. I spent some time figuring out if there was a
loop or not, but once I did I managed to get the correct answer in an
okish time. Once we know the size of the loop, we can skip all those
iterations, until we are very close to the target.

Part one: 00:08:54
Part two: 00:15:54