2017 Day 17
So this one I have no idea how to solve part two fast! It will be
interesting to see in the solutions thread!

My current solution for part two is slow, but it works when used with
pypy. I've tried to switch it to lists and to a linked list, but deque
has the absolute fastest access times. I made a mistake the first time,
checking for a different value than 0. Once I corrected that, I got the
right answer.

Part one: 00:04:17
Part two: 00:14:35