2017 Day 20
This was a mess, an absolute mess. Part one went quite quickly, but part
two was a major fuck up. I first tried to bruteforce it, but only found
one collision.

I then tried to devise an analytical method using the normal equations
of motion from physics. I did this using sympy. This turns out to be
quite slow and on top of that the formulas don't apply to this simulated
universe. The proper formula for the position of a vector at any given
time is:

r(t) = r_0 + v_0 * t + 1/2 * a * t * (t+1)

instead of the usual:

r(t) = r_0 + v_0 * t + 1/2 * a * t^2

Once I figured this out, it was still very slow, but I started seeing
some collisions in the very beginning of the program. This was weird to
me, since I hadn't seen this in my bruteforce.

I looked back on it and found out that I wasn't properly updating the Z
axis of the particles. Ones I did this, the bruteforce worked quite
quickly and I got the answer. My time on part two could have been
something like 10 minutes. ARGH!

At least I got to refresh my memory on the ABC formula!

Part one: 00:17:44
Part two: 02:02:23