2017 Day 21
Again we have something which uses a Conways Game of Life ish ruleset to
generate art. I've encountered these in later years, and so I quickly
reached for numpy. I was able to find a routine online to break the
array into subparts and "np.block" gave me a means to put them back
together.

I realized also that it would be faster to generate all permutations of
the input patterns and store them as strings. This could be hashed and
so would allow for an O(1) lookup once the building started. There is
probably some way to do this without holding everything in memory, but
for me this was good enough. Will be exciting to see how other people
did it.

PS: I'm very glad I did this in numpy as it allowed me to quickly rotate
the arrays and so on.

Part one: 00:33:25
Part two: 00:01:05