2017 Day 22
Funny one! It's a variation of langtons ants. I was a bit foolish, I had
a bug for a while, which was that I was using the normal Y increases
downward when initializing the board, but then used Y increases upwards
when doing the directions. Once I rectified this, I go the correct
answer.

I liked my transition to part two, I was able to get the answer quite
quickly there.

Part one: 00:21:53
Part two: 00:05:43