2017 Day 23
Now here was a challenge! Part one was simple enough, but part two was
really nice!

I began by trying to just run my normal program. This didn't work, so I
did what I have done before, which is to translate it into C and see if
my compiler would allow me to skip most of it. This didn't work and I
had a bunch of bugs in my translation, because of statements like:

sub d -1

Which really mean "d++", but I wrote as "d--". Once this was done, I
started running it and I got it running, the program was going up, but
it was going very slowly.

I then started reverse engineering the algorithm. I figured out in the
end that it tried to find all primes between two numbers X and Y, with
the step count of 17. For me, this can be solved like the following:

len([x for x in range(105700, 122701, 17) if not isprime(x)])

Where "isprime" is from sympy. Once I realized this, I could modify the
algorithm to not test all combinations, but rather only until the
product of the two variables was larger than the intended target.

This makes thing quicker, and it finished!

Really liked this task, be interesting to see if it's possible to make
it generic.

Part one: 00:04:02
Part two: 01:12:51