2017 Day 25
Feel like I should have been able to complete this one quicker, but
spent quite some time parsing the input. For the real challenge I'll be
better off just hardcoding the answer first and doing the parsing later.
This was about practice anyway, so for now it's ok that I spent the time
doing it properly.

The problem itself is quite trivial, but the parsing is the real trick
here.

I feel like this year was very good. Not as good as 2019, but still a
very good year.

Part one: 00:15:44