2021 Day 2
Well, today we learned the hard way that you need to trust the examples
and read the damn text. Part one went ok, had a bit of nerves, I should
have just gone for it and been more aggressive, but I'm ok with the
time. I ended up interpreting the directions the wrong way, so up
actually became down, but that just meant the sign on my answer was the
wrong way. I just ignored this and inverted it when I sent in the
answer, instead of trying to fix it before sending in.

In part two it fell apart though. I quite quickly got the aim variable
in, but I didn't get the correct score on the example. I should have
taken a few seconds to think about that, but instead I thought the
example might be a bit different this time, and so I tried anyway. I
submitted a wrong answer, and the game was up at that point. Took me way
to long to realize that now the depth is not updated on going up or
down.

You live you learn, but getting nerves this early ain't good ;)

Score:
      -------Part 1--------   -------Part 2--------
Day       Time  Rank  Score       Time  Rank  Score
  2   00:01:47   158      0   00:05:41  1272      0