2018 Day 6
Very cool day! It's basically an integer version of a Voronoi diagram! I
didn't solve it in a fancy way, but rather just calculated it for what
was on screen and just assume that anything touching the edges is going
to be infinite.

I'm ok with my time for part one, but it would have been a lot better if
I hadn't tripped over the fact that there was 50 points in the input set
and I zipped the input with "string.ascii_uppercase", which of course
only has 26 characters. Once I realized this, I quickly got the right
answer.

Part two, I just winged it and got an OK time.

I think this one will be a pain to optimize, so I'll leave that for
later!

Part one: 00:55:37
Part two: 00:08:35