2018 Day 7
This was an interesting problem! I remembered the "TopologicalSorter" in
pythons standard library "graphlib", which is exactly what is needed
here. I got thrown for a loop with the "alphabetical" requirement
however, which made me implement it by hand. I got it after a bit of
tinkering.

Part two was harder and I didn't really know how I was going to attack
it. I tried to do it the "smart" way, but I couldn't figure it out what
I needed here. So I ended up using a little trick! I inserted 10000 into
the answer box and submitted it. The helpful website will tell me that
this answer is too high, and so now I know that even a bruteforce will
be quite quick. So I end up just coding up this.

Questions for further consideration are if there is a very fast way to
do this or if I have to simulate each time step. I'm also looking closer
at the graphlib implementation, because it seems that I might be able to
use it's methods to bruteforce more efficiently.

Either way a good day!

Part one: 00:12:18
Part two: 00:34:26