2018 Day 8
An ok day, I felt slow on part one and ok fast on part two. Lost some
time on part two, due to not realizing that the meta data was 1 indexed
instead of 0 indexed.

I was mentally preparing for having to cache the node values in classic
dynamic programming style, but in hindsight this is obviously not
needed, as we ever only visit each node once.

Will clean this up in a second.

Part one: 00:06:47
Part two: 00:02:55