2018 Day 11
Spent a bit of time understanding the task, but once I understood it, it
was really just quite straight forward.

Part two I did a dirty tactic and just started a bruteforce search. Once
it had started, I started printing out the new best scores. Ones the
best score had stood for a while, I copied I tried entering it and,
tada, it was the right answer.

Looking at it now though, it is very obvious that we can use a summation
table here! I recently used this at work for something unrelated, but I
didn't think of using it during the contest. This was ok for now, as
doing it the slow way yielded a result quickly, but when I clean up I'll
be doing it correctly.

Part one: 00:09:52
Part two: 00:03:48