Cleanup of 2018 Day 11
So I was right and part two can be implemented effectively with a
summed-area table. This is not so important for part one, but for part
two it allows us to finish very quickly.

The solution is still a bit slow with normal python, but with pypy it is
nearly instant. I have a feeling that I could do better than this, but
that will have to be a task for a later date.