2018 Day 12
This was an ok day. It's another cellular automata task, but it part two
is spicy. I realized my machine was not going to be able to hack it. I
started printing the length of the array and saw that it didn't expand
after 67.

The difference in sum is also 67. This means that the robot is probably
just in a stable pattern, and moving one forward at a time. I hacked up
some quick logic for this, but it is not generic. I will attempt to come
up with something generic later.

Part one: 00:11:22
Part two: 00:06:06