2018 Day 13
Huh, this was an ok day I guess? It took a lot of time to code part one,
given all the possible turns and so on, but I got through it. Struggled
a little while, because I didn't account for how many ways you could
enter a turn and also I used modulo 2 instead of modulo 3 on the
generation step, which made it so that I had to visualize the steps to
see what was going wrong. Thank God that the puzzle included an
illustrated example!

For part two I adapted pretty quickly. It is pretty cool to think about,
that I would have gotten a 37th place globally if I had played this day
when it was going on!

Part one: 00:27:07
Part two: 00:06:12