2021 Day 7
So this was a disgrace, I am shamed by my performance here. This day was
easy to code, but I spent a lot of time in the begining thinking that we
where trying to sort the crabs, instead of just moving them to one
position. This is a MUCH easier problem and one I solved almost
instantly.

THEN I MADE ANOTHER BIG MISTAKE. I thought they wanted the position with
least sum, but rather, they wanted the sum. Had to wait a whole minute
for this *facepalm*.

Part two was easy, I just brute forced it at first, but looking at it
now, it is clear that you can use the triangle numbers. It would be
interesting to see if there is some way to get the best position or
close to it, with some sort of formula.

Score:
      --------Part 1--------   -------Part 2--------
Day       Time   Rank  Score       Time  Rank  Score
  7   00:08:07   3423      0   00:09:12  1480      0