2021 Day 8
Finally a good task! I did poorly in terms of time, but it was a task
that required thinking and logical exclusion! I spent quite a bit of
time before developing the tricks needed to get this, but after I
realized that the leaderboard position was not happening, I took my time
and enjoyed the processes.

The real trick here is to recognize that if two segments form pairs,
then if they are both present in the display, this narrows down the
possibilities. Once you get that down, it should resolve it self.

I might have been better of using z3 here, but I had started doing it in
pure python and I didn't want to convert. I'll do it later today, to
test out how this would look and to get antiquated with z3 again.

Overall, very happy with the task today, and feeling OK about my time.

Score:
      --------Part 1--------   -------Part 2--------
Day       Time   Rank  Score       Time  Rank  Score
  8   00:06:39    650      0   01:24:50  3815      0